Added jsdoc-compatible docstrings to source module; renamed lib source extensions to .JS (no react utilized at this time); added documentation artifacts ('public/') to gitignore; added jsdoc configuration file; incremented semver; added docs task to package scripts; added jsdoc and foodoc to development dependencies; expanded readme documentation to include package scripts; updated webpack configuration to reflect new/simplified file extensions
