+++ .gitignore
@@ -35,6 +35,7 @@ build/Release
 # Dependency directories
 node_modules/
 jspm_packages/
+package-lock.json
 
 # Typescript v1 declaration files
 typings/
@@ -57,3 +58,5 @@ typings/
 # dotenv environment variables file
 .env
 
+# documentation artifacts
+public/
