+++ package.json
@@ -26,7 +26,8 @@
   "main": "dist/rst2html.min.js",
   "scripts": {
     "test": "jest",
-    "build": "webpack"
+    "build": "webpack",
+    "docs": "node ./node_modules/jsdoc/jsdoc.js -d public/ -R ./README.md -c ./.jsdoc-conf.json -t ./node_modules/foodoc/template -r ."
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
@@ -34,7 +35,9 @@
     "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "jest": "^21.2.1",
-    "webpack": "^3.10.0"
+    "webpack": "^3.10.0",
+    "jsdoc": "^3.6.6",
+    "foodoc": "^0.0.9"
   },
   "dependencies": {
     "restructured": "0.0.11"
